refactor: use non-deprecated GraphQL Repository.name (not uri)
The Repository.uri field in GraphQL (and in general) is deprecated in favor of Repository.name. This is just a different name for the same value (`name` is a better name than `uri` because the value is not actually a URI, so `uri` is a misnomer). The value that both fields contain is the string of the form `github.com/foo/bar` for a repository.

The GraphQL `Query.repository` resolver also has 2 equivalent args, `name` and `uri`, and this also changes calls to use the former (for the same reason).